+++ front/src/components/Forms/index.ts
@@ -1,0 +1,1 @@
+export { FormUserProfile } from '@/components/Forms/FormUserProfile'
