+++ front/src/hooks/form/index.ts
@@ -1,0 +1,1 @@
+export { useCreateUpdateUserProfile } from '@/hooks/form/useCreateUpdateUserProfile'
